# Move Tracker to use direct sql queries