### All Submissions:

* [x] Have you followed the [WooCommerce Contributing guideline](https://github.com/woocommerce/woocommerce/blob/trunk/.github/CONTRIBUTING.md)?
* [x] Does your code follow the [WordPress' coding standards](https://make.wordpress.org/core/handbook/best-practices/coding-standards/)?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../pulls) for the same update/change?

<!-- Mark completed items with an [x] -->

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### Changes proposed in this Pull Request:

1. Closes #29839 by removing CRUD calls from WC Tracker code. This speeds up order analysis on larger stores.
2. Changes how order count by payment method is reported.

### How to test the changes in this Pull Request:

To compare speed or output to other versions of WooCommerce:
1. modify `WC_Tracker::get_orders()` to make it public
2. run `wp eval '$start = microtime(true);print_r(WC_Tracker::get_orders());echo microtime(true) - $start . " seconds";'`

Then do the same with another version. This only checks the modified section of code. To do this for the entire snapshot building process, do the above for `::get_tracking_data()` instead of `::get_orders()`.

### Other information:

* [x] Have you added an explanation of what your changes do and why you'd like us to include them?
* [ ] Have you written new tests for your changes, as applicable?
* [ ] Have you successfully run tests with your changes locally?

<!-- Mark completed items with an [x] -->

### Changelog entry

* Dev - Refactored Tracker to use direct DB calls instead of CRUD. #29877