@deltaWhiskey 2021-05-11
Checking the performance of `WC_Tracker::get_orders()` on a test store with 28,608 orders, 100 products, and 1,000 customers:
* the version in this PR finished in 0.005 to 0.008 seconds
* WooCommerce 5.1.1 (with CRUD functions) ran for 12-24 seconds, then exited with an out of memory error, having used just under 1 GB RAM
* WooCommerce 5.0.0 (before the CRUD functions were added) finished in 0.005 to 0.007 seconds.