@deltaWhiskey 2021-05-11
The WC Tracker snapshot generated by this PR was compared to that generated by WooCommerce 5.2.2, and was mostly the same. The differences seem acceptable to me. They were:
* This restores a pre-5.1.0 behavior when reporting order counts by status; this includes order statuses that have zero orders, while 5.2.2 omits them.
* This restores a pre-5.1.0 behavior for reporting certain money amounts; it uses a floating point number instead of limiting the precision.
* Version 5.2.2 reported order breakouts by gateway with keys like `gateway_paypal`. This version reports them like `gateway_paypal_USD_count` and `gateway_paypal_USD_total`.