@deltaWhiskey 2021-05-12
Thanks for the review, @roykho !

Removing `get_order_totals()` looks safe to me. It was made public in #22807 , but then that code was removed in #23302 .

Coding standard fixes have been applied.